backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
nltk>=3.6.3
textblob>=0.15.3
scikit-learn>=1.0.2
textstat>=0.7.2
pillow>=8.3.2
numpy>=1.21.2
opencv-python>=4.5.3.56
requests>=2.26.0 